Split Transaction into ExternalTransaction and Transfer
